Moved comparison functions to new module
